add methods to help the loaded_programs() API
Add helper methods to get useful information from the ProgramInfo
object which is returned by the `loaded_programs()` API.  Specifically
this code mirrors the bpftool program in terms of useful fields.

The methods here are not exhaustive and may expand in the future.

Add a new API macro to each aya `Program` type to allow us to fetch it's
accompanying `ProgramInfo` metadata after it's been loaded.

Add a new ProgramInfo constructor that builds a new instance using
a raw fd.

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>